Updated links storage
Instead of array we use a Map to store links on a global level

Nodes store links as set now.

Primary motivation for this change is that on
links heavy graphs removing a `X -> Y` link is
expensive:

`O(deg(X) + deg(Y))` in time space. With the
set change we now have `O(1)` performance.

It does affect microbenchmark for iterating over
sets/maps, but for now willing to take that hit.